Fixing test failure /w callbacks on 32 bit platforms

+++ hawtjni-runtime/src/main/java/org/fusesource/hawtjni/runtime/Callback.java
@@ -100,7 +100,7 @@ public class Callback {
      * @param errorResult
      *            the return value if the java code throws an exception
      */
-    public Callback(Object object, String method, int argCount, boolean isArrayBased, int /* long */errorResult) {
+    public Callback(Object object, String method, int argCount, boolean isArrayBased, long /* int */errorResult) {
 
         /* Set the callback fields */
         this.object = object;
